Move the drafter and lead selection into leads-core
The comment and DM drafter is where the product's judgement lives: its
hard rules are not style preferences but answers to the ways drafts
failed in real use — replying in the brand's language instead of the
thread's, filler openers, answering the title instead of the question in
the body, echoing what the top comments already said. That belongs with
the lead core, not inside a 1700-line orchestration module.

Moved: buildEngagePrompt with its schema, selectTopComments and its
relevance floor, and authorProfileUrl — which sat in radar.ts but
answers the same question as platformOf, where do I reach this person.
Their tests move with them; radar.ts imports them back.

SOURCES_SCHEMA and VERDICT_SCHEMA deliberately stay: their prompts are
built inline inside radarScan, and moving a schema away from the prompt
that fills it would be half a contract.

Adds the package README (the schema contract a reuser needs) and the
internal changelog. No public changelog: nothing here is visible from
outside.
